regular_worker: simplified selection of work to do
The new single query locks and returns an Operation ready to
be executed (all its inputs are in the 'present' state).
In itself, this would be interesting to upstream to AWB, but
it should also be noted that relying instead on the reservations
should instead provide a simpler way to do, more in line with our
slogans (we're doing it for the planner only now). This is more
complicated than for concrete applications, because the regular
worker would not be daemon for most of them, and instead would
react to operator requests.